refactor: Rename app/models.py to app/schemas.py for Pydantic models
- Move GitHubUserData from app/auth.py to app/schemas.py
- Keeps DB models in app/database/models.py, API schemas in app/schemas.py
